all the features
every tool now supports a toml config file. there are bad readmes for everything.

TLS is now used. the pertinent tools have config options for paths to certs.
as long as the cert is included everywhere it's fine that no trust store
validates it.

OPAQUE is now used for authentication of database edits. i hope to have a PAM
module that uses this functionality someday.